+++ CHANGELOG.md
@@ -13,6 +13,11 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - Removed third argument of `TTT2CanOrderEquipment`-Hook, no message is outputted anymore
 ### Fixed
 
+### Breaking Changes
+
+- Moved global shared `EquipmentIsBuyable(tbl, ply)` to `shop.CanBuyEquipment(ply, equipmentId)`
+  - Returned text and result are now replaced by a statusCode
+
 ## [v0.12.3b](https://github.com/TTT-2/TTT2/tree/v0.12.3b) (2024-01-07)
 
 ### Added